+++ CHANGES.rst
@@ -5,6 +5,9 @@
 0.16 (unreleased)
 =================
 
+- Add ``name`` and ``caption`` options. Added in
+  `PR 41 <https://github.com/NextThought/sphinxcontrib-programoutput/pull/41>`_ 
+  by Raphaël.
 - Add support for Python 3.8.